# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reloop API
  version: 1.0.0
  description: |
    API oficial del backend de Reloop.
    Incluye autenticación JWT, publicaciones, ofertas, carrito,
    compras, favoritos, ratings, usuarios y comentarios.
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #############################
  # GLOBAL SCHEMAS
  #############################
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, example: vale@example.com }
        username: { type: string }
        name: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        country: { type: string, nullable: true }
        birthDate:
          type: string
          format: date
          nullable: true
        avatar:
          type: string
          example: https://example.com/avatar.png
        createdAt:
          type: string
          format: date-time
...
    AuthSignupRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
        country:
          type: string
    AuthLoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          example: vale@example.com
        password:
          type: string
          example: "123456"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ###########################
    # LISTINGS
    ###########################
    ListingPhoto:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
    Listing:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        condition: { type: string }
        category: { type: string }
        subCategory: { type: string, nullable: true }
        subSubCategory: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        color: { type: string, nullable: true }
        seller:
          $ref: "#/components/schemas/User"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ListingPhoto"
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        createdAt:
          type: string
          format: date-time
        author:
          type: object
          properties:
            id: { type: string }
            username: { type: string }
            avatar: { type: string }
...
    FavoriteItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        price: { type: number }
        category: { type: string }
        condition: { type: string }
        photo: { type: string }
        seller:
          $ref: "#/components/schemas/User"
        favoriteId: { type: string }
...
    Rating:
      type: object
      properties:
        id: { type: string }
        value: { type: number }
        comment: { type: string }
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/User"
        listing:
          $ref: "#/components/schemas/Listing"
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Registro de usuario
      description: Crear un nuevo usuario mediante email, username y password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSignupRequest"
      responses:
        "201":
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email o username ya registrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login con email o username
      description: Autentica al usuario mediante email/username + password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/google:
    post:
      tags: [Auth]
      summary: Login/Signup con Google OAuth
      description: Autentica mediante token de Google y crea usuario si no existe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  example: ya29.a0AfH6SM123abc
      responses:
        "200":
          description: Autenticación correcta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Token faltante o inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  ###############################################
  # USERS
  ###############################################
...
  /comments/{listingId}:
    get:
      tags: [Comments]
      summary: Obtener comentarios públicos de un listing
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          description: Listing no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Comments]
      summary: Crear comentario público en un listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  example: "¿Qué tal es la calidad?"
      responses:
        "201":
          description: Comentario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Token inválido o faltante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  ###############################################
  # LISTINGS
  ###############################################
...
  /favorites:
    get:
      tags: [Favorites]
      summary: Listar favoritos del usuario autenticado
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest]
            default: newest
      responses:
        "200":
          description: Lista paginada de favoritos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FavoriteItem"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
        "401":
          description: Token inválido o faltante
...
  /cart:
    get:
      tags: [Cart]
      summary: Obtener items del carrito del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de items del carrito
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    userId: { type: string }
                    listingId: { type: string }
                    createdAt:
                      type: string
                      format: date-time
                    listing:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        price: { type: number }
                        photos:
                          type: array
                          items:
                            type: object
                            properties:
                              url: { type: string }
        "401":
          description: Token inválido o faltante
  /cart/{listingId}:
    post:
      tags: [Cart]
      summary: Agregar un item al carrito
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Item agregado o ya existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  userId: { type: string }
                  listingId: { type: string }
                  createdAt:
                    type: string
                    format: date-time
        "404":
          description: listing_not_found
        "400":
          description: cannot_add_own_listing_to_cart
        "401":
          description: Token inválido o faltante

    delete:
      tags: [Cart]
      summary: Quitar un item del carrito
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "401":
          description: Token inválido o faltante
  ###############################################
  # OFFERS
  ###############################################
...
  /ratings:
    post:
      tags: [Ratings]
      summary: Crear rating para un usuario
      description: Un usuario autenticado puede calificar a otro. No se permite auto-calificación.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetId, value]
              properties:
                targetId:
                  type: string
                  example: cl_user123
                value:
                  type: number
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  nullable: true
                  example: "Muy buena comunicación!"
                listingId:
                  type: string
                  nullable: true
                  example: cl_listing123
      responses:
        "200":
          description: Rating creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "400":
          description: Error de validación o auto-calificación
        "404":
          description: Usuario a calificar no existe
        "401":
          description: Token inválido o faltante
...